move the MySQL code into a seperate file
